# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivateResponse": {
        "properties": {
          "activated_id": {
            "title": "Activated Id",
            "type": "integer"
          },
          "activated_name": {
            "title": "Activated Name",
            "type": "string"
          },
          "previously_active": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previously Active"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "activated_id",
          "activated_name"
        ],
        "title": "ActivateResponse",
        "type": "object"
      },
...
      "Body_refresh_portfolio_api_v1_portfolios__portfolio_id__refresh_post": {
        "properties": {
          "column_mapping": {
            "description": "JSON: canonical_field \u2192 original_column_name",
            "title": "Column Mapping",
            "type": "string"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file",
          "column_mapping"
        ],
        "title": "Body_refresh_portfolio_api_v1_portfolios__portfolio_id__refresh_post",
        "type": "object"
      },
...
      "DeleteResponse": {
        "properties": {
          "deleted_id": {
            "title": "Deleted Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "deleted_id",
          "message"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoldingBase": {
        "properties": {
          "asset_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "Equity",
            "example": "Equity",
            "title": "Asset Class"
          },
          "average_cost": {
            "example": 3500.0,
            "exclusiveMinimum": 0.0,
            "title": "Average Cost",
            "type": "number"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "INR",
            "example": "INR",
            "title": "Currency"
          },
          "current_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "example": 3820.0,
            "title": "Current Price"
          },
          "data_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": "live",
            "title": "Data Source"
          },
          "enrichment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Overall enrichment quality: enriched|partial|failed|pending",
            "title": "Enrichment Status"
          },
          "fundamentals_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether fundamentals were fetched at upload: fetched|unavailable|pending",
            "title": "Fundamentals Status"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": "IT Services & Consulting",
            "title": "Industry"
          },
          "name": {
            "example": "Tata Consultancy Services",
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-text user annotations from the uploaded file",
            "title": "Notes"
          },
          "purchase_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Date of purchase (YYYY-MM-DD or as found in the file)",
            "example": "2023-04-15",
            "title": "Purchase Date"
          },
          "quantity": {
            "example": 30,
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "type": "number"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": "Information Technology",
            "title": "Sector"
          },
          "sector_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source that resolved this sector: from_file|yfinance|fmp|static_map|unknown",
            "title": "Sector Status"
          },
          "ticker": {
            "description": "Stock ticker symbol",
            "example": "TCS.NS",
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "name",
          "quantity",
          "average_cost"
        ],
        "title": "HoldingBase",
        "type": "object"
      },
...
      "PortfolioCreateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "source": {
            "default": "manual",
            "description": "manual | mock",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "PortfolioCreateRequest",
        "type": "object"
      },
...
      "PortfolioListResponse": {
        "properties": {
          "active_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Id"
          },
          "portfolios": {
            "items": {
              "$ref": "#/components/schemas/PortfolioMeta"
            },
            "title": "Portfolios",
            "type": "array"
          }
        },
        "required": [
          "portfolios"
        ],
        "title": "PortfolioListResponse",
        "type": "object"
      },
      "PortfolioMeta": {
        "description": "Lightweight portfolio summary \u2014 no holdings included.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_refreshable": {
            "description": "True for sources where the user can re-import/sync data.",
            "readOnly": true,
            "title": "Is Refreshable",
            "type": "boolean"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "num_holdings": {
            "default": 0,
            "title": "Num Holdings",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_meta_parsed": {
            "additionalProperties": true,
            "description": "Convenience: source_metadata as a parsed dict (empty if null/invalid).",
            "readOnly": true,
            "title": "Source Meta Parsed",
            "type": "object"
          },
          "source_metadata": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Metadata"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "upload_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Filename"
          }
        },
        "required": [
          "id",
          "name",
          "source",
          "is_active",
          "created_at",
          "updated_at",
          "is_refreshable",
          "source_meta_parsed"
        ],
        "title": "PortfolioMeta",
        "type": "object"
      },
...
      "PortfolioRenameRequest": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "PortfolioRenameRequest",
        "type": "object"
      },
...
      "RefreshResponse": {
        "description": "Returned by POST /portfolios/{id}/refresh \u2014 re-import into an existing portfolio.",
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "holdings_parsed": {
            "title": "Holdings Parsed",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "post_refresh_snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Refresh Snapshot Id"
          },
          "pre_refresh_snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pre Refresh Snapshot Id"
          },
          "rows_skipped": {
            "title": "Rows Skipped",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "portfolio_id",
          "filename",
          "holdings_parsed",
          "rows_skipped",
          "message"
        ],
        "title": "RefreshResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "P-Insight Portfolio Analytics API. Visit /docs for interactive API documentation.",
    "title": "P-Insight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/portfolio/": {
      "get": {
        "description": "Return all portfolio holdings from the active data source.",
        "operationId": "get_holdings_api_v1_portfolio__get",
        "parameters": [
          {
            "description": "Data source mode",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "uploaded",
              "description": "Data source mode",
              "enum": [
                "uploaded",
                "live",
                "broker"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HoldingBase"
                  },
                  "title": "Response Get Holdings Api V1 Portfolio  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get all holdings",
        "tags": [
          "Portfolio"
        ]
      }
    },
...
    "/api/v1/portfolios/": {
      "get": {
        "description": "Return all saved portfolios and indicate which is currently active.",
        "operationId": "list_portfolios_api_v1_portfolios__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List all portfolios",
        "tags": [
          "Portfolios"
        ]
      },
      "post": {
        "description": "Create a new empty portfolio with a given name.",
        "operationId": "create_portfolio_api_v1_portfolios__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a manual portfolio",
        "tags": [
          "Portfolios"
        ]
      }
    },
...
    "/api/v1/portfolios/{portfolio_id}": {
      "delete": {
        "operationId": "delete_portfolio_api_v1_portfolios__portfolio_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a portfolio",
        "tags": [
          "Portfolios"
        ]
      },
      "get": {
        "operationId": "get_portfolio_api_v1_portfolios__portfolio_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a portfolio",
        "tags": [
          "Portfolios"
        ]
      }
    },
    "/api/v1/portfolios/{portfolio_id}/activate": {
      "post": {
        "description": "Set the specified portfolio as active, deactivating all others.",
        "operationId": "activate_portfolio_api_v1_portfolios__portfolio_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate a portfolio",
        "tags": [
          "Portfolios"
        ]
      }
    },
...
    "/api/v1/portfolios/{portfolio_id}/refresh": {
      "post": {
        "description": "Re-import a new CSV/Excel file into an existing uploaded portfolio.\n\nThis preserves history by creating a pre- and post-refresh snapshot,\nthen replaces all holdings with the new file's data.\n\nArchitecture note: this is the same as /upload/confirm but targets an\nexisting portfolio rather than creating a new one \u2014 designed to support\nfuture broker-sync refresh flows.",
        "operationId": "refresh_portfolio_api_v1_portfolios__portfolio_id__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_refresh_portfolio_api_v1_portfolios__portfolio_id__refresh_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Re-import holdings into an existing portfolio",
        "tags": [
          "Portfolios"
        ]
      }
    },
    "/api/v1/portfolios/{portfolio_id}/rename": {
      "patch": {
        "operationId": "rename_portfolio_api_v1_portfolios__portfolio_id__rename_patch",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename a portfolio",
        "tags": [
          "Portfolios"
        ]
      }
    },
...
  }
}